fix(api): rename /remittance/qoute -> /remittance/quote with compat alias
Adds the correctly spelled route, keeps a deprecated alias re-export, and structures the fallback error response.
